Post 작성자만 삭제할 수 있도록 변경, login 수정

+++ app.py
@@ -15,7 +15,7 @@ def home():
     if request.method == 'POST':
         username = request.form.get('inputText')
     else:
-        username = session['username']
+        username = session['username'] 
 
     flash('hello, {}'.format(username))
     return render_template("home.html")
@@ -54,6 +54,7 @@ def post():
     posttable = Post()
     posttable.title = title
     posttable.content = content
+    posttable.username = session['username']  # post가 작성자 username에 대한 정보를 포함하도록 변경
 
     db.session.add(posttable)
     db.session.commit()
@@ -67,9 +68,12 @@ def detail(post_id):
 @app.route('/delete/<int:post_id>/')
 def delete(post_id):
     post = Post.query.get_or_404(post_id)
-    db.session.delete(post)
-    db.session.commit()
-    return redirect('/post_list/')
+    if post.username == session['username']:  # post의 작성자와 로그인된 사용자가 같을때만 delete 실행
+        db.session.delete(post)
+        db.session.commit()
+        return redirect('/post_list/')
+    else:
+        return "게시글은 작성자만 삭제할 수 있습니다"
 
 @app.route('/detail/<int:post_id>/comment/', methods=['GET', 'POST'])
 def comment(post_id):
@@ -96,7 +100,7 @@ def signup():
     username = request.form.get('username')
     password = request.form.get('password')
 
-    if not(username and password):
+    if not(username):
         return "사용자 이름이 입력되지 않았습니다"
 
     usertable=User()
@@ -119,13 +123,16 @@ def login():
 
     if not(username and password):
         return "입력되지 않은 정보가 있습니다"
-
-    db = sqlite3.connect("db.sqlite")
-    user = db.cursor()
-    user.execute("SELECT * FROM user WHERE username = '%s'" % username)
-    session['username'] = username
-    rows = user.fetchall()
-    return rows
+    else: # login 부분에서 비밀번호가 틀려도 로그인되고, [username, password]를 반환하던 점 수정
+        user = User.query.filter_by(username=username).first()
+        if user:
+            if user.password == password:
+                session['username'] = username
+                return redirect('/')
+            else:
+                return "비밀번호가 다릅니다"
+        else:
+            return "사용자가 존재하지 않습니다"
 
 @app.route('/logout/', methods=['GET', 'POST'])
 def logout():
